module2-task1 - Добавляет сборку проекта и розбивку на модули
