+++ public/index.html
@@ -47,5 +47,5 @@
       <!-- Контент -->
     </main>
   </body>
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </html>
