+++ src/components/load-more-button.js
@@ -1,0 +1,9 @@
+const createLoadMoreTemplate = () => {
+  return (
+    `
+    <button class="load-more" type="button">load more</button>
+    `
+  );
+};
+
+export {createLoadMoreTemplate};
